+++ src/ui/components/Account/LoginPage.css
@@ -1,0 +1,18 @@
+.login-page {
+  width: 100%;
+  height: 100%;
+}
+
+.login-page button {
+  background: #1890ff;
+  border: none;
+  color: white;
+  font-size: 16px;
+  border-radius: 8px;
+  padding: 16px;
+}
+
+.login-page button:hover {
+  cursor: pointer;
+  background-color: #2a90ff;
+}
