--- src/ui/components/Account/index.js
@@ -1,40 +1,0 @@
-import React, { useEffect } from "react";
-import useAuth0 from "ui/utils/useAuth0";
-import { setUserInBrowserPrefs } from "../../utils/browser";
-import Library from "./Library";
-
-import "./Account.css";
-
-function WelcomePage() {
-  const { loginWithRedirect } = useAuth0();
-  const forceOpenAuth = new URLSearchParams(window.location.search).get("signin");
-
-  if (forceOpenAuth) {
-    loginWithRedirect();
-    return null;
-  }
-
-  useEffect(() => {
-    setUserInBrowserPrefs(null);
-  }, []);
-
-  return (
-    <div className="welcome-screen">
-      <div className="welcome-panel">
-        <img className="logo" src="images/logo.svg" />
-        <img className="atwork" src="images/computer-work.svg" />
-        <button onClick={() => loginWithRedirect()}>Sign In</button>
-      </div>
-    </div>
-  );
-}
-
-export default function Account() {
-  const { isAuthenticated } = useAuth0();
-
-  if (!isAuthenticated) {
-    return <WelcomePage />;
-  }
-
-  return <Library />;
-}
